Fix analyzer issues

+++ src/PoshArmDeployment/Private/Remove-InternalProperty.ps1
@@ -1,21 +1,23 @@
 function Remove-InternalProperty {
-    [cmdletbinding()]
+    [cmdletbinding(SupportsShouldProcess = $True)]
     param(
         [Parameter(Mandatory, ValueFromPipeline)]
         [PSCustomObject]
         $InputObject
     )
 
-    $propNames = $($inputObject | Get-Member -MemberType *Property).Name
-    foreach ($propName in $propNames) {
-        if($propName.StartsWith("_")){
-            $InputObject.PSObject.Properties.Remove($propName)
+    If ($PSCmdlet.ShouldProcess("Removing all properties starting with an underscore on target object")) {
+        $propNames = $($inputObject | Get-Member -MemberType *Property).Name
+        foreach ($propName in $propNames) {
+            if ($propName.StartsWith("_")) {
+                $InputObject.PSObject.Properties.Remove($propName)
+            }
+            elseif ($InputObject.$propName.GetType().Name -eq "PSCustomObject") {
+                # Recurse
+                $InputObject.$propName = Remove-InternalProperty -InputObject $InputObject.$propName
+            }
         }
-        elseif ($InputObject.$propName.GetType().Name -eq "PSCustomObject") {
-            # Recurse
-            $InputObject.$propName = Remove-InternalProperty -InputObject $InputObject.$propName
-        }
-    }
 
-    return $InputObject
+        return $InputObject
+    }
 }
